Extract FlamechartPanZoomView to its own file
